Validate database copy moved
